# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: QA Work Management API
  version: 0.3.0
  description: |-
    Generated from the Zod schemas in packages/shared by scripts/openapi/generate.ts. Do not edit by hand.
    Errors are RFC 9457 problem details (docs/decisions/ADR-0010-problem-details-errors.md).
servers:
  - url: http://localhost:3000
    description: npm run dev
security:
  - session: []
...
paths:
...
  /api/projects:
    get:
      operationId: API-PROJECT-01
      summary: The caller's projects
      description: 'Docs: docs/api/projects/get-projects.md'
      tags:
        - Projects
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: archived
          in: query
          required: false
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectSummary'
        '400':
          description: 'Bad Request: VALIDATION_ERROR'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
    post:
      operationId: API-PROJECT-02
      summary: Create a project
      description: 'Docs: docs/api/projects/post-project.md'
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: URL of the new resource
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          description: 'Bad Request: VALIDATION_ERROR'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '409':
          description: 'Conflict: KEY_TAKEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - KEY_TAKEN
        '415':
          description: 'Unsupported Media Type: UNSUPPORTED_MEDIA_TYPE'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNSUPPORTED_MEDIA_TYPE
  /api/projects/{key}:
    get:
      operationId: API-PROJECT-03
      summary: One project
      description: 'Docs: docs/api/projects/get-project.md'
      tags:
        - Projects
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
    patch:
      operationId: API-PROJECT-04
      summary: Edit a project
      description: 'Docs: docs/api/projects/patch-project.md'
      tags:
        - Projects
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          description: 'Bad Request: VALIDATION_ERROR'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '409':
          description: 'Conflict: VERSION_CONFLICT'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - VERSION_CONFLICT
        '415':
          description: 'Unsupported Media Type: UNSUPPORTED_MEDIA_TYPE'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNSUPPORTED_MEDIA_TYPE
        '422':
          description: 'Unprocessable Content: PROJECT_ARCHIVED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - PROJECT_ARCHIVED
    delete:
      operationId: API-PROJECT-07
      summary: Delete an archived project
      description: 'Docs: docs/api/projects/delete-project.md'
      tags:
        - Projects
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '422':
          description: 'Unprocessable Content: DELETE_NOT_ALLOWED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - DELETE_NOT_ALLOWED
  /api/projects/{key}/archive:
    post:
      operationId: API-PROJECT-05
      summary: Archive a project
      description: 'Docs: docs/api/projects/post-project-archive.md'
      tags:
        - Projects
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '415':
          description: 'Unsupported Media Type: UNSUPPORTED_MEDIA_TYPE'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNSUPPORTED_MEDIA_TYPE
        '422':
          description: 'Unprocessable Content: PROJECT_ARCHIVED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - PROJECT_ARCHIVED
  /api/projects/{key}/restore:
    post:
      operationId: API-PROJECT-06
      summary: Restore a project
      description: 'Docs: docs/api/projects/post-project-restore.md'
      tags:
        - Projects
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '415':
          description: 'Unsupported Media Type: UNSUPPORTED_MEDIA_TYPE'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNSUPPORTED_MEDIA_TYPE
  /api/projects/{key}/members:
    get:
      operationId: API-PROJECT-08
      summary: Members of a project
      description: 'Docs: docs/api/projects/get-project-members.md'
      tags:
        - Members
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Member'
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
    post:
      operationId: API-PROJECT-09
      summary: Add a member
      description: 'Docs: docs/api/projects/post-project-member.md'
      tags:
        - Members
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberAdd'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Member'
        '400':
          description: 'Bad Request: VALIDATION_ERROR'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '409':
          description: 'Conflict: ALREADY_MEMBER'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - ALREADY_MEMBER
        '415':
          description: 'Unsupported Media Type: UNSUPPORTED_MEDIA_TYPE'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNSUPPORTED_MEDIA_TYPE
        '422':
          description: 'Unprocessable Content: PROJECT_ARCHIVED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - PROJECT_ARCHIVED
  /api/projects/{key}/members/{userId}:
    patch:
      operationId: API-PROJECT-10
      summary: Change a member's access level or job title
      description: 'Docs: docs/api/projects/patch-project-member.md'
      tags:
        - Members
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User id of the member
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Member'
        '400':
          description: 'Bad Request: VALIDATION_ERROR'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '415':
          description: 'Unsupported Media Type: UNSUPPORTED_MEDIA_TYPE'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNSUPPORTED_MEDIA_TYPE
        '422':
          description: 'Unprocessable Content: PROJECT_ARCHIVED, OWN_ACCESS, LAST_PROJECT_ADMIN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - PROJECT_ARCHIVED
                          - OWN_ACCESS
                          - LAST_PROJECT_ADMIN
    delete:
      operationId: API-PROJECT-11
      summary: Remove a member, or leave
      description: 'Docs: docs/api/projects/delete-project-member.md'
      tags:
        - Members
      parameters:
        - name: key
          in: path
          required: true
          description: Project key, case-insensitive
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User id of the member
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '401':
          description: 'Unauthorized: UNAUTHENTICATED'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - UNAUTHENTICATED
        '403':
          description: 'Forbidden: FORBIDDEN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - FORBIDDEN
        '404':
          description: 'Not Found: NOT_FOUND'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - NOT_FOUND
        '422':
          description: 'Unprocessable Content: PROJECT_ARCHIVED, LAST_PROJECT_ADMIN'
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - properties:
                      code:
                        enum:
                          - PROJECT_ARCHIVED
                          - LAST_PROJECT_ADMIN
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: qawm_sid
      description: Set by POST /api/auth/login
  schemas:
...
    ProjectSummary:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        archivedAt:
          anyOf:
            - type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
            - type: 'null'
        myAccess:
          anyOf:
            - type: string
              enum:
                - PROJECT_ADMIN
                - MEMBER
                - GUEST
            - type: 'null'
        memberCount:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        activeRelease:
          anyOf:
            - type: object
              properties:
                id:
                  type: string
                name:
                  type: string
              required:
                - id
                - name
              additionalProperties: false
            - type: 'null'
        updatedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$
      required:
        - key
        - name
        - archivedAt
        - myAccess
        - memberCount
        - activeRelease
        - updatedAt
      additionalProperties: false
...
    ProjectCreate:
      type: object
      properties:
        key:
          type: string
          minLength: 1
        firstAdminId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          anyOf:
            - type: string
              maxLength: 2000
            - type: 'null'
      required:
        - key
        - firstAdminId
        - name
      additionalProperties: false
    ProjectUpdate:
      type: object
      properties:
        version:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          anyOf:
            - type: string
              maxLength: 2000
            - type: 'null'
      required:
        - version
      additionalProperties: false
    EmptyBody:
      type: object
      properties: {}
      additionalProperties: false
    MemberAdd:
      type: object
      properties:
        userId:
          type: string
          minLength: 1
        access:
          type: string
          enum:
            - PROJECT_ADMIN
            - MEMBER
            - GUEST
        jobTitle:
          anyOf:
            - type: string
              enum:
                - QAE
                - QAL
                - QAA
                - PM
                - PO
                - BA
                - DEV
                - TL
                - DES
                - STK
                - OTH
            - type: 'null'
      required:
        - userId
        - access
      additionalProperties: false
...